jetson/tx2: boot to squashfs with mutable overlayfs
The old boot process is available by setting TX2_MUTABLE_ROOTFS when using the
flashusb script.

Previously, the tx2 booted to a mutable ext2 APP partition. Now, the tx2 will
use the APP partition as persist, with /boot containing the usual Skiff boot-up
files - squashfs and Image.

A custom C tool skiff-init-squashfs is added, which is executed by the kernel as
the init process with APP (persist) mounted to / as root. The init process will
mount the squashfs to /skiff, persist to /skiff/mnt/persist, and an overlayfs
over /skiff. Then, it chroots into /skiff and runs systemd.

Static resize2fs is optionally bundled with skiff-init, and executed
before mounting the rootfs image. It reads skiff-init/resize2fs.conf
ignoring comments starting with # for paths to /dev/ block devices to
resize. This resizes the persist partition before systemd starts.

Signed-off-by: Christian Stewart <christian@paral.in>