pine64/rockpro64: switch to skiff kernel 5.15.32
from greylinux on Discord:

All of them up to and including 5.15 worked with our any thermal errors , 5.16
is when the thermal error started, but all including 5.16 work with the pcie and
show the ssds.

5.17 has the thermal errors and the ssd's don't show up, even though I can see
the pcie with lspci, I'm not sure if its working properly because the drives
dont show.

    5.13 no issue
    5.13.19 no issue
    5.15 no issue
    5.16 thermal error
    5.17 thermal and PCIe issue

Rolling back kernel to 5.15.x version for now.

See issue #190

Signed-off-by: Christian Stewart <christian@paral.in>