skiff-init-squashfs: mount / upper dir as tmpfs
The overlayfs code now checks if the lower dir changed. Previously we re-used the same upper dir, which gave a persistent / overlay, however, this was never the desired behavior: SkiffOS uses a mutable / with a tmpfs, but changes should be wiped between boots.

Mount a tmpfs to the upper dir of the overlay instead of a persistent dir

Signed-off-by: Christian Stewart <christian@paral.in>